cluster-setup/requirements.txt
aws-cdk-lib==2.47.0
constructs>=10.0.0
